moved incidents endpoints into its own router
